Add files via upload
The input file of CNF model in terms of COVID-19 recoveries, including eight columns and 32 rows, where the first seven columns are related to single natural factors (SNFs), and the last column is about pandemic variable, i.e., daily recovery growth rate. The purpose of CNF model is assigning seven weight values to each SNFs to make the correlation between CNF (The outcome of linear combination of seven SNFs with corresponding weights) and pandemic variable as high as possible.